pyproject.toml
[project]
name = "deepagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "deepagents>=0.3.7",
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-groq>=1.1.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "semantic-chunkers>=0.1.1",
    "semantic-router>=0.1.12",
    "sentence-transformers>=5.2.0",
    "tavily-python>=0.7.19",
    "torch>=2.10.0",
    "trafilatura>=2.0.0",
    "transformers>=4.51.0",
    "uvicorn>=0.40.0",
]
